Allow reading of XA_WINDOW properties
This opens up some cool possibilities, e.g. checking if a window is
transient for another one:

    local xid    = tostring(get_window_xid())
    local tf_xid = get_window_property('WM_TRANSIENT_FOR')
    if tf_xid ~= '' then
        debug_print('The window ' .. xid .. ' is kept above ' .. tf_xid)
    end

The only problem I see is get_window_xid() returns a number while
get_window_property() returns a string so, if you need to make
comparations between them, you need to cast the former to a string or
the latter to a number (in Lua "123" is *not* equal to 123).